Separate keyboard input from player input
