--- client.js
@@ -1,1042 +1,0 @@
-(function e(t,n,r){function s(o,u){if(!n[o]){if(!t[o]){var a=typeof require=="function"&&require;if(!u&&a)return a(o,!0);if(i)return i(o,!0);var f=new Error("Cannot find module '"+o+"'");throw f.code="MODULE_NOT_FOUND",f}var l=n[o]={exports:{}};t[o][0].call(l.exports,function(e){var n=t[o][1][e];return s(n?n:e)},l,l.exports,e,t,n,r)}return n[o].exports}var i=typeof require=="function"&&require;for(var o=0;o<r.length;o++)s(r[o]);return s})({1:[function(require,module,exports){
-'use strict';
-
-Object.defineProperty(exports, '__esModule', {
-  value: true
-});
-var dogeImage = new Image();
-dogeImage.src = 'doge.png';
-
-exports['default'] = dogeImage;
-module.exports = exports['default'];
-
-},{}],2:[function(require,module,exports){
-"use strict";
-
-Object.defineProperty(exports, "__esModule", {
-  value: true
-});
-
-var _createClass = (function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ("value" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; })();
-
-function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
-
-var EventStream = (function () {
-  function EventStream() {
-    _classCallCheck(this, EventStream);
-
-    this.listeners = {};
-  }
-
-  _createClass(EventStream, [{
-    key: "on",
-    value: function on(eventName, callback) {
-      this.listeners[eventName] = this.listeners[eventName] || [];
-      this.listeners[eventName].push(callback);
-    }
-  }, {
-    key: "broadcast",
-    value: function broadcast(eventName) {
-      for (var _len = arguments.length, data = Array(_len > 1 ? _len - 1 : 0), _key = 1; _key < _len; _key++) {
-        data[_key - 1] = arguments[_key];
-      }
-
-      (this.listeners[eventName] || []).forEach(function (callback) {
-        callback.call.apply(callback, [null].concat(data));
-      });
-    }
-  }]);
-
-  return EventStream;
-})();
-
-exports["default"] = EventStream;
-module.exports = exports["default"];
-
-},{}],3:[function(require,module,exports){
-'use strict';
-
-var _bind = Function.prototype.bind;
-
-var _createClass = (function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ('value' in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; })();
-
-function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { 'default': obj }; }
-
-function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError('Cannot call a class as a function'); } }
-
-var _msgpackJsBrowser = require('msgpack-js-browser');
-
-var _msgpackJsBrowser2 = _interopRequireDefault(_msgpackJsBrowser);
-
-var _inputListener = require('./input-listener');
-
-var _inputListener2 = _interopRequireDefault(_inputListener);
-
-var _rendering = require('./rendering');
-
-var _rendering2 = _interopRequireDefault(_rendering);
-
-var _stage = require('./stage');
-
-var _stage2 = _interopRequireDefault(_stage);
-
-window.Client = (function () {
-  function Client(name) {
-    _classCallCheck(this, Client);
-
-    this.name = name;
-  }
-
-  _createClass(Client, [{
-    key: 'start',
-    value: function start() {
-      var stage = new _stage2['default'](1000, 600);
-      var input = new _inputListener2['default']();
-      var client = new WebSocket(document.location.protocol.replace('http', 'ws') + '//' + document.location.host);
-      var thrusties = [];
-
-      window.addEventListener('beforeunload', function () {
-        client.close();
-      });
-
-      client.binaryType = 'arraybuffer';
-
-      client.onopen = (function () {
-        client.send(_msgpackJsBrowser2['default'].encode({ join: { name: this.name } }));
-
-        input.events.on('stateChange', function (state) {
-          client.send(_msgpackJsBrowser2['default'].encode({ inputState: state }));
-        });
-      }).bind(this);
-
-      client.onmessage = function (message) {
-        var gameState = _msgpackJsBrowser2['default'].decode(message.data).state;
-        new _rendering2['default'](stage, gameState, thrusties).perform();
-      };
-    }
-  }], [{
-    key: 'start',
-    value: function start() {
-      for (var _len = arguments.length, params = Array(_len), _key = 0; _key < _len; _key++) {
-        params[_key] = arguments[_key];
-      }
-
-      var client = new (_bind.apply(Client, [null].concat(params)))();
-      client.start();
-    }
-  }]);
-
-  return Client;
-})();
-
-},{"./input-listener":4,"./rendering":5,"./stage":6,"msgpack-js-browser":7}],4:[function(require,module,exports){
-'use strict';
-
-Object.defineProperty(exports, '__esModule', {
-  value: true
-});
-
-var _createClass = (function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ('value' in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; })();
-
-function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { 'default': obj }; }
-
-function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError('Cannot call a class as a function'); } }
-
-var _eventStream = require('./event-stream');
-
-var _eventStream2 = _interopRequireDefault(_eventStream);
-
-var InputListener = (function () {
-  function InputListener() {
-    _classCallCheck(this, InputListener);
-
-    this.events = new _eventStream2['default']();
-    this.state = {};
-    this.bindToEvents();
-  }
-
-  _createClass(InputListener, [{
-    key: 'keyCodeAction',
-    value: function keyCodeAction(keyCode) {
-      var keyCodeToActionMappings = {
-        38: 'thrust',
-        37: 'left',
-        39: 'right'
-      };
-
-      return keyCodeToActionMappings[keyCode];
-    }
-  }, {
-    key: 'bindToEvents',
-    value: function bindToEvents() {
-      window.addEventListener('keydown', this.setState(true));
-      window.addEventListener('keyup', this.setState(false));
-    }
-  }, {
-    key: 'setState',
-    value: function setState(value) {
-      return (function (event) {
-        var action = this.keyCodeAction(event.keyCode);
-
-        if (action) {
-          this.state[action] = value;
-        }
-
-        this.events.broadcast('stateChange', this.state);
-      }).bind(this);
-    }
-  }]);
-
-  return InputListener;
-})();
-
-exports['default'] = InputListener;
-module.exports = exports['default'];
-
-},{"./event-stream":2}],5:[function(require,module,exports){
-'use strict';
-
-Object.defineProperty(exports, '__esModule', {
-  value: true
-});
-
-var _slicedToArray = (function () { function sliceIterator(arr, i) { var _arr = []; var _n = true; var _d = false; var _e = undefined; try { for (var _i = arr[Symbol.iterator](), _s; !(_n = (_s = _i.next()).done); _n = true) { _arr.push(_s.value); if (i && _arr.length === i) break; } } catch (err) { _d = true; _e = err; } finally { try { if (!_n && _i['return']) _i['return'](); } finally { if (_d) throw _e; } } return _arr; } return function (arr, i) { if (Array.isArray(arr)) { return arr; } else if (Symbol.iterator in Object(arr)) { return sliceIterator(arr, i); } else { throw new TypeError('Invalid attempt to destructure non-iterable instance'); } }; })();
-
-var _createClass = (function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ('value' in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; })();
-
-function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { 'default': obj }; }
-
-function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError('Cannot call a class as a function'); } }
-
-var _doge = require('./doge');
-
-var _doge2 = _interopRequireDefault(_doge);
-
-var Rendering = (function () {
-  function Rendering(stage, gameState, thrusties) {
-    _classCallCheck(this, Rendering);
-
-    this.stage = stage;
-    this.context = stage.context;
-    this.gameState = gameState;
-    this.thrusties = thrusties;
-  }
-
-  _createClass(Rendering, [{
-    key: 'perform',
-    value: function perform() {
-      this.erase();
-      this.drawBackground();
-      this.drawThrusties();
-      this.drawShips();
-    }
-  }, {
-    key: 'erase',
-    value: function erase() {
-      this.draw(function (context, stage) {
-        context.clearRect(0, 0, stage.width, stage.height);
-      });
-    }
-  }, {
-    key: 'drawThrusties',
-    value: function drawThrusties() {
-      var _this = this;
-
-      this.draw(function (context) {
-        _this.thrusties.forEach(function (_ref, index) {
-          var angle = _ref.angle;
-          var x = _ref.x;
-          var y = _ref.y;
-          var mx = _ref.mx;
-          var my = _ref.my;
-          var ttl = _ref.ttl;
-
-          _this.thrusties[index].ttl -= 0.3;
-          _this.thrusties[index].x += mx;
-          _this.thrusties[index].y += my;
-          if (ttl > 0) {
-            context.fillStyle = 'rgba(225, 0, 0, ' + ttl + ')';
-            _this.drawTriangle(x, y, 10 * ttl, angle);
-            context.fill();
-          } else {
-            _this.thrusties.splice(index, 1);
-          }
-        });
-      });
-    }
-  }, {
-    key: 'drawShips',
-    value: function drawShips() {
-      var _this2 = this;
-
-      this.gameState.ships.forEach(function (ship) {
-        _this2.draw(function (context) {
-
-          if (ship.doge) {
-            this.drawDoge(ship);
-          } else {
-            context.fillStyle = ship.colour;
-            this.drawTriangle(ship.x, ship.y, 30, ship.rotation);
-            context.fill();
-          }
-
-          this.drawText(ship.name, ship.x, Math.max(ship.y - 25, 10), ship.colour);
-
-          if (ship.thrusting) {
-            var _rotatePoint = this.rotatePoint(-ship.rotation, 0, 0)([0, 10]);
-
-            var _rotatePoint2 = _slicedToArray(_rotatePoint, 2);
-
-            var mx = _rotatePoint2[0];
-            var my = _rotatePoint2[1];
-
-            this.thrusties.push({
-              angle: ship.rotation,
-              x: ship.x + mx,
-              y: ship.y + my,
-              mx: mx,
-              my: my,
-              ttl: 1
-            });
-          }
-        });
-      });
-    }
-  }, {
-    key: 'drawDoge',
-    value: function drawDoge(ship) {
-      this.drawImage(_doge2['default'], ship.x, ship.y, 96, 51, ship.rotation);
-    }
-
-    // Draw image whose center is x, y
-  }, {
-    key: 'drawImage',
-    value: function drawImage(image, x, y, width, height) {
-      var rotation = arguments.length <= 5 || arguments[5] === undefined ? 0 : arguments[5];
-
-      this.draw(function (context) {
-        context.translate(x, y);
-        context.rotate(-rotation);
-        context.drawImage(image, -width / 2, -height / 2, width, height);
-      });
-    }
-
-    // Draw text whose center is at x
-  }, {
-    key: 'drawText',
-    value: function drawText(text, x, y, colour) {
-      this.draw(function (context) {
-        var measurement = context.measureText(text);
-        context.fillStyle = colour;
-        context.font = "bold 10pt Helvetica Neue";
-        context.fillText(text, x - measurement.width / 2, y);
-      });
-    }
-  }, {
-    key: 'drawBackground',
-    value: function drawBackground() {
-      this.draw(function (context, stage) {
-        context.fillStyle = 'white';
-        context.fillRect(0, 0, stage.width, stage.height);
-      });
-    }
-  }, {
-    key: 'rotatePoint',
-    value: function rotatePoint(angle, x, y) {
-      var cos = Math.cos(angle);
-      var sin = Math.sin(angle);
-      return function (_ref2) {
-        var _ref22 = _slicedToArray(_ref2, 2);
-
-        var px = _ref22[0];
-        var py = _ref22[1];
-        return [px * cos - py * sin + x, px * sin + py * cos + y];
-      };
-    }
-  }, {
-    key: 'drawTriangle',
-    value: function drawTriangle(x, y, size, rotation) {
-      var path = [[0, -size / 2], [size / 2, size / 2], [-size / 2, size / 2]].map(this.rotatePoint(-rotation, x, y));
-
-      this.drawPath(path);
-    }
-  }, {
-    key: 'drawPath',
-    value: function drawPath(points) {
-      this.draw(function (context) {
-        context.beginPath();
-        context.moveTo.apply(context, points.shift());
-        points.forEach(function (point) {
-          return context.lineTo.apply(context, point);
-        });
-      });
-    }
-  }, {
-    key: 'draw',
-    value: function draw(callback) {
-      this.context.save();
-      callback.call(this, this.context, this.stage);
-      this.context.restore();
-    }
-  }]);
-
-  return Rendering;
-})();
-
-exports['default'] = Rendering;
-module.exports = exports['default'];
-
-},{"./doge":1}],6:[function(require,module,exports){
-'use strict';
-
-Object.defineProperty(exports, '__esModule', {
-  value: true
-});
-
-var _createClass = (function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ('value' in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; })();
-
-function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError('Cannot call a class as a function'); } }
-
-var Stage = (function () {
-  function Stage(width, height) {
-    _classCallCheck(this, Stage);
-
-    this.materialize();
-    this.width = width;
-    this.height = height;
-  }
-
-  _createClass(Stage, [{
-    key: 'materialize',
-    value: function materialize() {
-      this.canvas = document.createElement('canvas');
-      this.context = this.canvas.getContext('2d');
-      document.body.appendChild(this.canvas);
-    }
-  }, {
-    key: 'width',
-    set: function set(value) {
-      this.canvas.width = value;
-    },
-    get: function get() {
-      return this.canvas.width;
-    }
-  }, {
-    key: 'height',
-    set: function set(value) {
-      this.canvas.height = value;
-    },
-    get: function get() {
-      return this.canvas.height;
-    }
-  }]);
-
-  return Stage;
-})();
-
-exports['default'] = Stage;
-module.exports = exports['default'];
-
-},{}],7:[function(require,module,exports){
-( // Module boilerplate to support browser globals and browserify and AMD.
-  typeof define === "function" ? function (m) { define("msgpack-js", m); } :
-  typeof exports === "object" ? function (m) { module.exports = m(); } :
-  function(m){ this.msgpack = m(); }
-)(function () {
-"use strict";
-
-var exports = {};
-
-exports.inspect = inspect;
-function inspect(buffer) {
-  if (buffer === undefined) return "undefined";
-  var view;
-  var type;
-  if (buffer instanceof ArrayBuffer) {
-    type = "ArrayBuffer";
-    view = new DataView(buffer);
-  }
-  else if (buffer instanceof DataView) {
-    type = "DataView";
-    view = buffer;
-  }
-  if (!view) return JSON.stringify(buffer);
-  var bytes = [];
-  for (var i = 0; i < buffer.byteLength; i++) {
-    if (i > 20) {
-      bytes.push("...");
-      break;
-    }
-    var byte = view.getUint8(i).toString(16);
-    if (byte.length === 1) byte = "0" + byte;
-    bytes.push(byte);
-  }
-  return "<" + type + " " + bytes.join(" ") + ">";
-}
-
-// Encode string as utf8 into dataview at offset
-exports.utf8Write = utf8Write;
-function utf8Write(view, offset, string) {
-  var byteLength = view.byteLength;
-  for(var i = 0, l = string.length; i < l; i++) {
-    var codePoint = string.charCodeAt(i);
-
-    // One byte of UTF-8
-    if (codePoint < 0x80) {
-      view.setUint8(offset++, codePoint >>> 0 & 0x7f | 0x00);
-      continue;
-    }
-
-    // Two bytes of UTF-8
-    if (codePoint < 0x800) {
-      view.setUint8(offset++, codePoint >>> 6 & 0x1f | 0xc0);
-      view.setUint8(offset++, codePoint >>> 0 & 0x3f | 0x80);
-      continue;
-    }
-
-    // Three bytes of UTF-8.  
-    if (codePoint < 0x10000) {
-      view.setUint8(offset++, codePoint >>> 12 & 0x0f | 0xe0);
-      view.setUint8(offset++, codePoint >>> 6  & 0x3f | 0x80);
-      view.setUint8(offset++, codePoint >>> 0  & 0x3f | 0x80);
-      continue;
-    }
-
-    // Four bytes of UTF-8
-    if (codePoint < 0x110000) {
-      view.setUint8(offset++, codePoint >>> 18 & 0x07 | 0xf0);
-      view.setUint8(offset++, codePoint >>> 12 & 0x3f | 0x80);
-      view.setUint8(offset++, codePoint >>> 6  & 0x3f | 0x80);
-      view.setUint8(offset++, codePoint >>> 0  & 0x3f | 0x80);
-      continue;
-    }
-    throw new Error("bad codepoint " + codePoint);
-  }
-}
-
-exports.utf8Read = utf8Read;
-function utf8Read(view, offset, length) {
-  var string = "";
-  for (var i = offset, end = offset + length; i < end; i++) {
-    var byte = view.getUint8(i);
-    // One byte character
-    if ((byte & 0x80) === 0x00) {
-      string += String.fromCharCode(byte);
-      continue;
-    }
-    // Two byte character
-    if ((byte & 0xe0) === 0xc0) {
-      string += String.fromCharCode(
-        ((byte & 0x0f) << 6) | 
-        (view.getUint8(++i) & 0x3f)
-      );
-      continue;
-    }
-    // Three byte character
-    if ((byte & 0xf0) === 0xe0) {
-      string += String.fromCharCode(
-        ((byte & 0x0f) << 12) |
-        ((view.getUint8(++i) & 0x3f) << 6) |
-        ((view.getUint8(++i) & 0x3f) << 0)
-      );
-      continue;
-    }
-    // Four byte character
-    if ((byte & 0xf8) === 0xf0) {
-      string += String.fromCharCode(
-        ((byte & 0x07) << 18) |
-        ((view.getUint8(++i) & 0x3f) << 12) |
-        ((view.getUint8(++i) & 0x3f) << 6) |
-        ((view.getUint8(++i) & 0x3f) << 0)
-      );
-      continue;
-    }
-    throw new Error("Invalid byte " + byte.toString(16));
-  }
-  return string;
-}
-
-exports.utf8ByteCount = utf8ByteCount;
-function utf8ByteCount(string) {
-  var count = 0;
-  for(var i = 0, l = string.length; i < l; i++) {
-    var codePoint = string.charCodeAt(i);
-    if (codePoint < 0x80) {
-      count += 1;
-      continue;
-    }
-    if (codePoint < 0x800) {
-      count += 2;
-      continue;
-    }
-    if (codePoint < 0x10000) {
-      count += 3;
-      continue;
-    }
-    if (codePoint < 0x110000) {
-      count += 4;
-      continue;
-    }
-    throw new Error("bad codepoint " + codePoint);
-  }
-  return count;
-}
-
-exports.encode = function (value) {
-  var buffer = new ArrayBuffer(sizeof(value));
-  var view = new DataView(buffer);
-  encode(value, view, 0);
-  return buffer;
-}
-
-exports.decode = decode;
-
-// http://wiki.msgpack.org/display/MSGPACK/Format+specification
-// I've extended the protocol to have two new types that were previously reserved.
-//   buffer 16  11011000  0xd8
-//   buffer 32  11011001  0xd9
-// These work just like raw16 and raw32 except they are node buffers instead of strings.
-//
-// Also I've added a type for `undefined`
-//   undefined  11000100  0xc4
-
-function Decoder(view, offset) {
-  this.offset = offset || 0;
-  this.view = view;
-}
-Decoder.prototype.map = function (length) {
-  var value = {};
-  for (var i = 0; i < length; i++) {
-    var key = this.parse();
-    value[key] = this.parse();
-  }
-  return value;
-};
-Decoder.prototype.buf = function (length) {
-  var value = new ArrayBuffer(length);
-  (new Uint8Array(value)).set(new Uint8Array(this.view.buffer, this.offset, length), 0);
-  this.offset += length;
-  return value;
-};
-Decoder.prototype.raw = function (length) {
-  var value = utf8Read(this.view, this.offset, length);
-  this.offset += length;
-  return value;
-};
-Decoder.prototype.array = function (length) {
-  var value = new Array(length);
-  for (var i = 0; i < length; i++) {
-    value[i] = this.parse();
-  }
-  return value;
-};
-Decoder.prototype.parse = function () {
-  var type = this.view.getUint8(this.offset);
-  var value, length;
-  // FixRaw
-  if ((type & 0xe0) === 0xa0) {
-    length = type & 0x1f;
-    this.offset++;
-    return this.raw(length);
-  }
-  // FixMap
-  if ((type & 0xf0) === 0x80) {
-    length = type & 0x0f;
-    this.offset++;
-    return this.map(length);
-  }
-  // FixArray
-  if ((type & 0xf0) === 0x90) {
-    length = type & 0x0f;
-    this.offset++;
-    return this.array(length);
-  }
-  // Positive FixNum
-  if ((type & 0x80) === 0x00) {
-    this.offset++;
-    return type;
-  }
-  // Negative Fixnum
-  if ((type & 0xe0) === 0xe0) {
-    value = this.view.getInt8(this.offset);
-    this.offset++;
-    return value;
-  }
-  switch (type) {
-  // raw 16
-  case 0xda:
-    length = this.view.getUint16(this.offset + 1);
-    this.offset += 3;
-    return this.raw(length);
-  // raw 32
-  case 0xdb:
-    length = this.view.getUint32(this.offset + 1);
-    this.offset += 5;
-    return this.raw(length);
-  // nil
-  case 0xc0:
-    this.offset++;
-    return null;
-  // false
-  case 0xc2:
-    this.offset++;
-    return false;
-  // true
-  case 0xc3:
-    this.offset++;
-    return true;
-  // undefined
-  case 0xc4:
-    this.offset++;
-    return undefined;
-  // uint8
-  case 0xcc:
-    value = this.view.getUint8(this.offset + 1);
-    this.offset += 2;
-    return value;
-  // uint 16
-  case 0xcd:
-    value = this.view.getUint16(this.offset + 1);
-    this.offset += 3;
-    return value;
-  // uint 32
-  case 0xce:
-    value = this.view.getUint32(this.offset + 1);
-    this.offset += 5;
-    return value;
-  // int 8
-  case 0xd0:
-    value = this.view.getInt8(this.offset + 1);
-    this.offset += 2;
-    return value;
-  // int 16
-  case 0xd1:
-    value = this.view.getInt16(this.offset + 1);
-    this.offset += 3;
-    return value;
-  // int 32
-  case 0xd2:
-    value = this.view.getInt32(this.offset + 1);
-    this.offset += 5;
-    return value;
-  // map 16
-  case 0xde:
-    length = this.view.getUint16(this.offset + 1);
-    this.offset += 3;
-    return this.map(length);
-  // map 32
-  case 0xdf:
-    length = this.view.getUint32(this.offset + 1);
-    this.offset += 5;
-    return this.map(length);
-  // array 16
-  case 0xdc:
-    length = this.view.getUint16(this.offset + 1);
-    this.offset += 3;
-    return this.array(length);
-  // array 32
-  case 0xdd:
-    length = this.view.getUint32(this.offset + 1);
-    this.offset += 5;
-    return this.array(length);
-  // buffer 16
-  case 0xd8:
-    length = this.view.getUint16(this.offset + 1);
-    this.offset += 3;
-    return this.buf(length);
-  // buffer 32
-  case 0xd9:
-    length = this.view.getUint32(this.offset + 1);
-    this.offset += 5;
-    return this.buf(length);
-  // float
-  case 0xca:
-    value = this.view.getFloat32(this.offset + 1);
-    this.offset += 5;
-    return value;
-  // double
-  case 0xcb:
-    value = this.view.getFloat64(this.offset + 1);
-    this.offset += 9;
-    return value;
-  }
-  throw new Error("Unknown type 0x" + type.toString(16));
-};
-function decode(buffer) {
-  var view = new DataView(buffer);
-  var decoder = new Decoder(view);
-  var value = decoder.parse();
-  if (decoder.offset !== buffer.byteLength) throw new Error((buffer.byteLength - decoder.offset) + " trailing bytes");
-  return value;
-}
-
-function encode(value, view, offset) {
-  var type = typeof value;
-
-  // Strings Bytes
-  if (type === "string") {
-    var length = utf8ByteCount(value);
-    // fix raw
-    if (length < 0x20) {
-      view.setUint8(offset, length | 0xa0);
-      utf8Write(view, offset + 1, value);
-      return 1 + length;
-    }
-    // raw 16
-    if (length < 0x10000) {
-      view.setUint8(offset, 0xda);
-      view.setUint16(offset + 1, length);
-      utf8Write(view, offset + 3, value);
-      return 3 + length;
-    }
-    // raw 32
-    if (length < 0x100000000) {
-      view.setUint8(offset, 0xdb);
-      view.setUint32(offset + 1, length);
-      utf8Write(view, offset + 5, value);
-      return 5 + length;
-    }
-  }
-
-  if (value instanceof ArrayBuffer) {
-    var length = value.byteLength;
-    // buffer 16
-    if (length < 0x10000) {
-      view.setUint8(offset, 0xd8);
-      view.setUint16(offset + 1, length);
-      (new Uint8Array(view.buffer)).set(new Uint8Array(value), offset + 3);
-      return 3 + length;
-    }
-    // buffer 32
-    if (length < 0x100000000) {
-      view.setUint8(offset, 0xd9);
-      view.setUint32(offset + 1, length);
-      (new Uint8Array(view.buffer)).set(new Uint8Array(value), offset + 5);
-      return 5 + length;
-    }
-  }
-  
-  if (type === "number") {
-    // Floating Point
-    if ((value << 0) !== value) {
-      view.setUint8(offset, 0xcb);
-      view.setFloat64(offset + 1, value);
-      return 9;
-    }
-
-    // Integers
-    if (value >=0) {
-      // positive fixnum
-      if (value < 0x80) {
-        view.setUint8(offset, value);
-        return 1;
-      }
-      // uint 8
-      if (value < 0x100) {
-        view.setUint8(offset, 0xcc);
-        view.setUint8(offset + 1, value);
-        return 2;
-      }
-      // uint 16
-      if (value < 0x10000) {
-        view.setUint8(offset, 0xcd);
-        view.setUint16(offset + 1, value);
-        return 3;
-      }
-      // uint 32
-      if (value < 0x100000000) {
-        view.setUint8(offset, 0xce);
-        view.setUint32(offset + 1, value);
-        return 5;
-      }
-      throw new Error("Number too big 0x" + value.toString(16));
-    }
-    // negative fixnum
-    if (value >= -0x20) {
-      view.setInt8(offset, value);
-      return 1;
-    }
-    // int 8
-    if (value >= -0x80) {
-      view.setUint8(offset, 0xd0);
-      view.setInt8(offset + 1, value);
-      return 2;
-    }
-    // int 16
-    if (value >= -0x8000) {
-      view.setUint8(offset, 0xd1);
-      view.setInt16(offset + 1, value);
-      return 3;
-    }
-    // int 32
-    if (value >= -0x80000000) {
-      view.setUint8(offset, 0xd2);
-      view.setInt32(offset + 1, value);
-      return 5;
-    }
-    throw new Error("Number too small -0x" + (-value).toString(16).substr(1));
-  }
-  
-  // undefined
-  if (type === "undefined") {
-    view.setUint8(offset, 0xc4);
-    return 1;
-  }
-  
-  // null
-  if (value === null) {
-    view.setUint8(offset, 0xc0);
-    return 1;
-  }
-
-  // Boolean
-  if (type === "boolean") {
-    view.setUint8(offset, value ? 0xc3 : 0xc2);
-    return 1;
-  }
-  
-  // Container Types
-  if (type === "object") {
-    var length, size = 0;
-    var isArray = Array.isArray(value);
-
-    if (isArray) {
-      length = value.length;
-    }
-    else {
-      var keys = Object.keys(value);
-      length = keys.length;
-    }
-
-    var size;
-    if (length < 0x10) {
-      view.setUint8(offset, length | (isArray ? 0x90 : 0x80));
-      size = 1;
-    }
-    else if (length < 0x10000) {
-      view.setUint8(offset, isArray ? 0xdc : 0xde);
-      view.setUint16(offset + 1, length);
-      size = 3;
-    }
-    else if (length < 0x100000000) {
-      view.setUint8(offset, isArray ? 0xdd : 0xdf);
-      view.setUint32(offset + 1, length);
-      size = 5;
-    }
-
-    if (isArray) {
-      for (var i = 0; i < length; i++) {
-        size += encode(value[i], view, offset + size);
-      }
-    }
-    else {
-      for (var i = 0; i < length; i++) {
-        var key = keys[i];
-        size += encode(key, view, offset + size);
-        size += encode(value[key], view, offset + size);
-      }
-    }
-    
-    return size;
-  }
-  throw new Error("Unknown type " + type);
-}
-
-function sizeof(value) {
-  var type = typeof value;
-
-  // Raw Bytes
-  if (type === "string") {
-    var length = utf8ByteCount(value);
-    if (length < 0x20) {
-      return 1 + length;
-    }
-    if (length < 0x10000) {
-      return 3 + length;
-    }
-    if (length < 0x100000000) {
-      return 5 + length;
-    }
-  }
-  
-  if (value instanceof ArrayBuffer) {
-    var length = value.byteLength;
-    if (length < 0x10000) {
-      return 3 + length;
-    }
-    if (length < 0x100000000) {
-      return 5 + length;
-    }
-  }
-  
-  if (type === "number") {
-    // Floating Point
-    // double
-    if (value << 0 !== value) return 9;
-
-    // Integers
-    if (value >=0) {
-      // positive fixnum
-      if (value < 0x80) return 1;
-      // uint 8
-      if (value < 0x100) return 2;
-      // uint 16
-      if (value < 0x10000) return 3;
-      // uint 32
-      if (value < 0x100000000) return 5;
-      // uint 64
-      if (value < 0x10000000000000000) return 9;
-      throw new Error("Number too big 0x" + value.toString(16));
-    }
-    // negative fixnum
-    if (value >= -0x20) return 1;
-    // int 8
-    if (value >= -0x80) return 2;
-    // int 16
-    if (value >= -0x8000) return 3;
-    // int 32
-    if (value >= -0x80000000) return 5;
-    // int 64
-    if (value >= -0x8000000000000000) return 9;
-    throw new Error("Number too small -0x" + value.toString(16).substr(1));
-  }
-  
-  // Boolean, null, undefined
-  if (type === "boolean" || type === "undefined" || value === null) return 1;
-  
-  // Container Types
-  if (type === "object") {
-    var length, size = 0;
-    if (Array.isArray(value)) {
-      length = value.length;
-      for (var i = 0; i < length; i++) {
-        size += sizeof(value[i]);
-      }
-    }
-    else {
-      var keys = Object.keys(value);
-      length = keys.length;
-      for (var i = 0; i < length; i++) {
-        var key = keys[i];
-        size += sizeof(key) + sizeof(value[key]);
-      }
-    }
-    if (length < 0x10) {
-      return 1 + size;
-    }
-    if (length < 0x10000) {
-      return 3 + size;
-    }
-    if (length < 0x100000000) {
-      return 5 + size;
-    }
-    throw new Error("Array or object too long 0x" + length.toString(16));
-  }
-  throw new Error("Unknown type " + type);
-}
-
-return exports;
-
-});
-
-},{}]},{},[3]);
