+++ lib/client/index.js
@@ -1,5 +1,6 @@
 import msgpack from 'msgpack-js-browser'
-import InputListener from './input-listener'
+import PlayerInput from './player-input'
+import KeyboardInput from './keyboard-input'
 import Rendering from './rendering'
 import Stage from './stage'
 
@@ -10,7 +11,7 @@ window.Client = class Client {
 
   start() {
     let stage = new Stage(1000, 600)
-    let input = new InputListener()
+    let input = new PlayerInput(KeyboardInput)
     let client = new WebSocket(document.location.protocol.replace('http', 'ws') + '//' + document.location.host)
     let thrusties = []
